Running unity.camera_client displays image. Moved frame_preview to a separate file. Fixed linter warning in rov_startup. Added comment in unity communication frame function explaining the transcoding.
