+++ unity/camera_client.py
@@ -5,6 +5,7 @@ import cv2
 from logpy.LogPy import Logger
 from definitions import LOG_DIRECOTRY
 from definitions import IP_ADDRESS, CAMERA_SERVER_PORT
+from utils.testing import frame_preview
 
 class CameraClient:
     def __init__(self, unity_reference, name_modifer = ""):
@@ -28,3 +29,11 @@ class CameraClient:
         frame = self.unity_reference.frame
         # data = data[msg_size:]
         return frame
+
+if __name__ == "__main__":
+    # shows one frame from the simulation
+    from unity.communication import Communication
+    communication=Communication()
+    camCl = CameraClient(communication)
+    frame_preview(camCl)
+    communication.close_simulation()
